Place reusable SplitPane component in common folder
