+++ src/common/components/ProgressCircle/ProgressCircle.js
