+++ src/common/components/ProgressCircle/ProgressCircle.module.scss
